pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epistemic-tribunal"
version = "0.1.0"
description = "Metacognitive reasoning architecture: Epistemic Tribunal for ARC-like tasks"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Epistemic Tribunal Contributors" }]
keywords = ["ai", "reasoning", "metacognition", "arc", "epistemic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "numpy>=1.26",
    "scipy>=1.11",
    "python-dotenv>=1.0",
    "transformers>=4.30",
    "openai>=1.0.0",
    "textual>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]
llm = [
    "accelerate>=0.20",
]

[project.scripts]
tribunal = "epistemic_tribunal.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/epistemic_tribunal"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py311"
